requirements.txt
# Core ML
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
Pillow==11.1.0
scikit-learn==1.6.1

# API
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20

# Experiment Tracking
mlflow==2.19.0

# Data Versioning
dvc==3.56.0

# Config
pyyaml==6.0.2

# Monitoring
prometheus-client==0.21.1

# Utilities
matplotlib==3.10.0
